facial-recognition/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
opencv-python==4.8.1.78
insightface==0.7.3
onnxruntime==1.16.3
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
